Pull EventEmitter into its own file
